+++ src/shared/sitemap.ts
@@ -1,10 +1,8 @@
 import Sitemapper from "sitemapper";
-import { fetchRetry } from "./utils.mjs";
+import { fetchRetry } from "./utils";
 
-async function getSitemapsList(accessToken, siteUrl) {
-  const url = `https://www.googleapis.com/webmasters/v3/sites/${encodeURIComponent(
-    siteUrl
-  )}/sitemaps`;
+async function getSitemapsList(accessToken: string, siteUrl: string) {
+  const url = `https://www.googleapis.com/webmasters/v3/sites/${encodeURIComponent(siteUrl)}/sitemaps`;
 
   const response = await fetchRetry(url, {
     headers: {
@@ -29,10 +27,10 @@ async function getSitemapsList(accessToken, siteUrl) {
   return body.sitemap.map((x) => x.path);
 }
 
-export async function getSitemapPages(accessToken, siteUrl) {
+export async function getSitemapPages(accessToken: string, siteUrl: string) {
   const sitemaps = await getSitemapsList(accessToken, siteUrl);
 
-  let pages = [];
+  let pages: string[] = [];
   for (const url of sitemaps) {
     const Google = new Sitemapper({
       url,
