+++ src/shared/utils.ts
@@ -1,7 +1,12 @@
-const createChunks = (arr, size) =>
-  Array.from({ length: Math.ceil(arr.length / size) }, (v, i) => arr.slice(i * size, i * size + size));
+const createChunks = (arr: any[], size: number) =>
+  Array.from({ length: Math.ceil(arr.length / size) }, (_, i) => arr.slice(i * size, i * size + size));
 
-export async function batch(task, items, batchSize, onBatchComplete) {
+export async function batch(
+  task: (url: string) => void,
+  items: string[],
+  batchSize: number,
+  onBatchComplete: (batchIndex: number, batchCount: number) => void
+) {
   const chunks = createChunks(items, batchSize);
   for (let i = 0; i < chunks.length; i++) {
     await Promise.all(chunks[i].map(task));
@@ -9,7 +14,7 @@ export async function batch(task, items, batchSize, onBatchComplete) {
   }
 }
 
-export async function fetchRetry(url, options, retries = 5) {
+export async function fetchRetry(url: string, options: RequestInit, retries: number = 5) {
   try {
     const response = await fetch(url, options);
     if (response.status >= 500) {
